requirements.txt
aiocontextvars==0.2.2
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
Click==7.0
colorclass==2.2.0
colorlog==4.1.0
contextvars==2.4
croniter==0.3.31
dataclasses==0.7
fastapi==0.52.0
faust==1.10.4
h11==0.9.0
httptools==0.1.1
idna==2.9
idna-ssl==1.1.0
immutables==0.11
kafka-python==2.0.1
mode==4.3.2
multidict==4.7.5
mypy-extensions==0.4.3
opentracing==1.3.0
pkg-resources==0.0.0
pydantic==1.4
python-dateutil==2.8.1
robinhood-aiokafka==1.1.6
six==1.14.0
starlette==0.13.2
terminaltables==3.1.0
typing-extensions==3.7.4.1
uvicorn==0.11.3
uvloop==0.14.0
venusian==1.2.0
websockets==8.1
yarl==1.4.2
